# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/roles": {
            "get": {
                "description": "查询所有/指定条件角色的信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "查询角色信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "roleCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "roleName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "新增不用传ID，修改才传ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "新增/修改角色信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "修改才需要传，没有传算新增",
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "roleCode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "roleDesc",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "roleName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            1,
                            2
                        ],
                        "type": "integer",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/roles/": {
            "delete": {
                "description": "删除指定角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "角色ID",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/roles/all-summary": {
            "get": {
                "description": "查询所有角色的简略信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "所有角色简略信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/roles/apis": {
            "get": {
                "description": "由于swagger本身的限制，get请求的切片会报错，并非接口本身问题，请换个方式，如http://127.0.0.1:9081/api/v1/group/apis?ids=3\u0026ids=4",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "获取角色的API权限列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 登录返回的用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/roles/bind": {
            "put": {
                "description": "1: api 2: menu 3: button",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "关联角色关系",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "associationType参数 1: api 2: menu 3: button",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BindRoleRelationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/roles/menus": {
            "get": {
                "description": "获取角色的关联菜单列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "获取角色的关联菜单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 登录返回的用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/roles/users": {
            "get": {
                "description": "获取角色的关联用户列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色相关"
                ],
                "summary": "获取角色的关联用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "格式为：Bearer 登录返回的用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":{},\"meta\":{msg\":\"Success\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"data\":{}, \"meta\":{\"msg\":\"错误信息\", \"error\":\"错误格式输出(如存在)\"}}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.BindRoleRelationReq": {
            "type": "object",
            "required": [
                "associationType",
                "objectIds",
                "roleId"
            ],
            "properties": {
                "associationType": {
                    "description": "1: api 2: menu 3: button",
                    "type": "integer"
                },
                "objectIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "roleId": {
                    "description": "角色id",
                    "type": "integer"
                }
            }
        },
...
        "api.IdsReq": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}